+++ src/mailcore/__init__.py
@@ -10,7 +10,7 @@ from mailcore.attachment import Attachment, AttachmentResolver, IMAPResolver, Si
 from mailcore.body import MessageBody
 from mailcore.draft import Draft
 from mailcore.email_address import EmailAddress
-from mailcore.exceptions import FolderNotFoundError, MailcoreError
+from mailcore.exceptions import FolderNotFoundError, MailcoreError, SMTPError
 from mailcore.folder import Folder
 from mailcore.mailbox import FolderDict, Mailbox
 from mailcore.message import Message
@@ -42,4 +42,5 @@ __all__ = [
     "SMTPConnection",
     "MailcoreError",
     "FolderNotFoundError",
+    "SMTPError",
 ]